\name{annotationGrangesList}
\alias{annotationGrangesList}
\docType{data}
\title{Annotation granges list prepared from bambu for the used chromosome}
\description{Annotation granges list prepared from bambu for the used chromosome}
\usage{
  data("annotationGrangesList")
}
\format{SummarizedExperiment}
\keyword{datasets}