\name{annotationGrangesList_chr22_1_25409234}
\alias{annotationGrangesList_chr22_1_25409234}
\docType{data}
\title{Annotation granges list prepared from bambu for the used chromosome}
\description{Annotation granges list prepared from bambu for the used chromosome}
\usage{
  data("annotationGrangesList_chr22_1_25409234")
}
\format{SummarizedExperiment}
\keyword{datasets}