\name{annotationGrangesList_chr22_1_34500000}
\alias{annotationGrangesList_chr22_1_34500000}
\docType{data}
\title{Annotation granges list prepared from bambu for the used chromosome}
\description{Annotation granges list prepared from bambu for the used chromosome}
\usage{
  data("annotationGrangesList_chr22_1_34500000")
}
\format{SummarizedExperiment}
\keyword{datasets}