\name{bamFiles}
\alias{bamFiles}
\docType{data}
\title{Bam file names}
\description{Describe available bam file names}
\details{
sample_info contains the following information:

\itemize{
\item source - The Singapore Nanopore Expression Project (SG-NEx)
\item reference genome - GRCh38
\item chromosome region - chr22:1-25409234
\item cell type - K562 (leukocyte) and MCF7 (breast)
\item name of files- K562_directRNA_replicate6.bam, K562_directcDNA_replicate1.bam, K562_directcDNA_replicate4.bam, MCF7_directRNA_replicate4.bam, MCF7_directcDNA_replicate1.bam, MCF7_directcDNA_replicate3.bam}

}
\usage{
  data("bamFiles")
}
\format{BamFileList}
\keyword{datasets}
