\name{HsChr22BambuAnnotation}
\alias{HsChr22BambuAnnotation}
\docType{data}
\title{BambuAnnotation of the first half of Human Sapiens Chr22}
\description{Annotation GRangesList prepared from bambu for Granges
of human genome (Grch38) chromosome 22 (1:25409234)}
\usage{
    data("HsChr22BambuAnnotation")
}
\format{SummarizedExperiment}
\keyword{datasets}
