\name{HsChr22HalfBambuAnnotation}
\alias{HsChr22HalfBambuAnnotation}
\docType{data}
\title{BambuAnnotation of the first half of Human Sapiens Chr22}
\description{Annotation GRangesList prepared from bambu for Granges of
the first half of Human Sapiens Chr22}
\usage{
    data("HsChr22HalfBambuAnnotation")
}
\format{SummarizedExperiment}
\keyword{datasets}
