\name{K562MCF7Samples}
\alias{K562MCF7Samples}
\docType{data}
\title{K562 and MCF7 samples from Nanopore RNA-Seq}
\description{Sample information description for K562 and MCF7 
samples from Nanopore RNA-Seq}
\details{
K562MCF7Samples is a DataFrame containing the following information:

\itemize{
\item sample_id - sample names of the bam files
\item Platform - sequencing platform
\item cellLine - cell line used
\item protocol - sequencing protocols
\item cancer_type- cancer type of the cell line}

}
\usage{
    data("K562MCF7Samples")
}
\format{DataFrame}
\keyword{datasets}
