\name{sample_info}
\alias{sample_info}
\docType{data}
\title{Sample information of the replicates}
\description{Sample information description for the two cancer cell lines}
\details{
sample_info is a DataFrame containing the following information:

\itemize{
\item sample_id - sample names of the bam files
\item Platform - sequencing platform
\item cellLine - cell line used
\item protocol - sequencing protocols
\item cancer_type- cancer type of the cell line}

}
\usage{
    data("sample_info")
}
\format{DataFrame}
\keyword{datasets}
