\name{SGNexSamples}
\alias{SGNexSamples}
\docType{data}
\title{SG-Nex samples from Nanopore RNA-Seq}
\description{Sample information description for K562 and MCF7 
samples from SG-Nex}
\details{
SGNexSamples is a DataFrame containing the following information:

\itemize{
\item sample_id - sample names of the bam files
\item Platform - sequencing platform
\item cellLine - cell line used
\item protocol - sequencing protocols
\item cancer_type- cancer type of the cell line}

}
\usage{
    data("SGNexSamples")
}
\format{DataFrame}
\keyword{datasets}
